fix: customer 정보 조회시 customer로 감싸서 반환하도록 수정
