+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import woowacourse.auth.application.AuthService;
+import woowacourse.auth.dto.CustomerDto;
 import woowacourse.auth.dto.CustomerResponse;
 import woowacourse.auth.dto.LoginRequest;
 import woowacourse.auth.dto.PasswordDto;
@@ -37,7 +38,7 @@ public class AuthController {
     @GetMapping
     public ResponseEntity<CustomerResponse> findCustomerInfo(HttpServletRequest request) {
         Customer customer = authService.findCustomerByUsername((String) request.getAttribute("username"));
-        return ResponseEntity.ok().body(CustomerResponse.from(customer));
+        return ResponseEntity.ok().body(new CustomerResponse(CustomerDto.from(customer)));
     }
 
     @PostMapping("/password")
